fix bug : 把03的数字类型读取成String类型，数据出错。

+++ src/main/java/cn/emay/excel/utils/ExcelReadUtils.java
@@ -249,7 +249,7 @@ public class ExcelReadUtils {
 		switch (ctype) {
 		case NUMERIC:
 			double d = cell.getNumericCellValue();
-			int dd = new Double(d).intValue();
+			long dd = new Double(d).longValue();
 			if (d == new Double(dd).doubleValue()) {
 				str = String.valueOf(dd);
 			} else {
